requirements.txt
flask
gunicorn
requests
beautifulsoup4
python-dotenv
playwright
